+++ tests/browser/conftest.py
@@ -10,7 +10,7 @@ from playwright.sync_api import Page
 from starlette.applications import Starlette
 from starlette.routing import Mount
 
-from .apps import make_app, make_events_app, make_lazy_app
+from .apps import make_app, make_events_app, make_lazy_app, make_live_app
 
 
 @pytest.fixture(scope="session")
@@ -19,6 +19,7 @@ def server_url() -> Iterator[str]:
         routes=[
             Mount("/lazy", app=make_lazy_app()),
             Mount("/events", app=make_events_app()),
+            Mount("/live", app=make_live_app()),
             Mount("/", app=make_app()),
         ]
     )